Speed up rod model (#28)
I have increased the rod model's speed by 50% by calculating the
lagrange bases only once. The problem is that each constraint in the
position level will call the "eval_basis" function to get the position
and orientation of the rod, so this function generates lagrange bases
every time and consumes time.